Move interleaveSplits() to util_math.h
