+++ test/unit/test_util_math.cpp
@@ -220,3 +220,47 @@ TYPED_TEST(MathUtilTest, ptrdiff_t_Arithmetic_SumMul) {
     EXPECT_EQ(expected_result, sum(mul(a, -1), c));
   }
 }
+
+TEST(Splits, DifferentOffsets) {
+  SizeType l = 10;
+  SizeType b = 3;
+  SizeType o1 = 1;
+  SizeType o2 = 2;
+  std::vector<SizeType> splits = util::interleaveSplits(l, b, o1, o2);
+
+  std::vector<SizeType> expected_splits{0, 1, 2, 4, 5, 7, 8, 10};
+  ASSERT_TRUE(splits == expected_splits);
+}
+
+TEST(Splits, EqualOffsets) {
+  SizeType l = 10;
+  SizeType b = 3;
+  SizeType o1 = 2;
+  SizeType o2 = 2;
+  std::vector<SizeType> splits = util::interleaveSplits(l, b, o1, o2);
+
+  std::vector<SizeType> expected_splits{0, 2, 5, 8, 10};
+  ASSERT_TRUE(splits == expected_splits);
+}
+
+TEST(Splits, SingleZeroOffset) {
+  SizeType l = 10;
+  SizeType b = 3;
+  SizeType o1 = 2;
+  SizeType o2 = 0;
+  std::vector<SizeType> splits = util::interleaveSplits(l, b, o1, o2);
+
+  std::vector<SizeType> expected_splits{0, 2, 3, 5, 6, 8, 9, 10};
+  ASSERT_TRUE(splits == expected_splits);
+}
+
+TEST(Splits, ZeroOffsets) {
+  SizeType l = 10;
+  SizeType b = 3;
+  SizeType o1 = 0;
+  SizeType o2 = 0;
+  std::vector<SizeType> splits = util::interleaveSplits(l, b, o1, o2);
+
+  std::vector<SizeType> expected_splits{0, 3, 6, 9, 10};
+  ASSERT_TRUE(splits == expected_splits);
+}
